LPS-97802 make DDMFormInstanceRecordUADUserCache a private class of DDMFormInstanceRecordUADDisplay
